ngrx - refactoring, code cleaning
