+++ courses-app/src/app/auth/guards/admin.guard.ts
@@ -2,20 +2,20 @@ import {Injectable} from '@angular/core';
 import {ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot, UrlTree} from '@angular/router';
 import {Observable} from 'rxjs';
 import {map} from "rxjs/operators";
-import {UserStoreService} from "../../user/user-store.service";
 import {UserService} from "../../user/user.service";
 
 @Injectable({
   providedIn: 'root'
 })
 export class AdminGuard implements CanActivate {
-  constructor(private userStoreService: UserStoreService, private router: Router, private userService: UserService) {
+  constructor(private router: Router, private userService: UserService) {
   }
 
   canActivate(
     route: ActivatedRouteSnapshot,
     state: RouterStateSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
 
+    // We are sending a request, because we would like to validate the token too
     return this.userService.fetchUser().pipe(
       map((user) => {
         if (user.role === 'admin') {
