+++ courses-app/src/app/features/courses/course-list/course-list.component.html
@@ -7,8 +7,3 @@
     </ng-container>
   </ng-container>
 </app-course-card>
-<p (click)="demoNgrx()" class="btn btn-success m-3">Click me - trigger Ngrx REQUEST_CURRENT_USER_START</p>
-<p> Loading? - {{ userState.$isLoading | async }}</p>
-<p> Username? - {{ userState.$name | async }}</p>
-<p> isAdmin? - {{ userState.$isAdmin | async }}</p>
-<p> role? - {{ userState.$role | async }}</p>
