+++ courses-app/src/app/shared/components/header/header.component.ts
@@ -1,6 +1,4 @@
 import {Component, OnInit} from '@angular/core';
-import {UserStoreService} from "../../../user/user-store.service";
-import {AuthService} from "../../../auth/services/auth.service";
 import {UserStateFacade} from "../../../user/store/user.facade";
 import {AuthStateFacade} from "../../../auth/store/auth.facade";
 
@@ -10,26 +8,12 @@ import {AuthStateFacade} from "../../../auth/store/auth.facade";
   styleUrls: ['./header.component.scss']
 })
 export class HeaderComponent implements OnInit {
-  userName: string | undefined;
-  userRole: string | undefined;
-  isAuthorized: boolean = false;
-
-
-  constructor(private userStoreService: UserStoreService,
-              public userStateFacade: UserStateFacade,
-              public authStateFacade: AuthStateFacade,
-              private authService: AuthService) {
+  constructor(
+    public userStateFacade: UserStateFacade,
+    public authStateFacade: AuthStateFacade) {
   }
 
   ngOnInit(): void {
-    this.userStoreService.user$.subscribe(user => {
-      this.userName = user?.name;
-      this.userRole = user?.role;
-    })
-
-    this.authService.isAuthorized$.subscribe(isAuthorized => {
-      this.isAuthorized = isAuthorized;
-    })
   }
 
   logout() {
